gestaltd: delete duplicate external credential sync path
Old interfaces:
- ExternalCredentialService lived in canonical_future.go next to IdentityDelegationService.
- TokenService had two mirror write paths: syncExternalCredential(...) for live writes and a separate syncExternalCredentialRecord(...) implementation for startup backfill.

New interfaces:
- ExternalCredentialService lives in external_credentials.go, while IdentityDelegationService stays in identity_delegations.go.
- TokenService still exposes the same two entry points, but startup backfill now reuses the live syncExternalCredential(...) path instead of reimplementing the upsert.

Examples:
- old startup backfill: integration token record -> decrypt -> rebuild ExternalCredential inline -> UpsertCredential
- new startup backfill: integration token record -> decrypt -> syncExternalCredential(...)
- old file layout: canonical_future.go mixed delegations and external credential mirror code
- new file layout: external_credentials.go and identity_delegations.go are split by responsibility

Notes:
- no runtime interface or schema change
- deletes duplicated canonical external credential assembly logic
